fix bug if u back to english from another

+++ start.py
@@ -1,47 +1,50 @@
-# VERSION: 2.0
+# VERSION: 2.1
 import os
 import re
-import sys
-import datetime
 
-folder_path = "C:\\Program Files (x86)\\Steam\\steamapps"
+folder_path = r"C:\Users\admin\Desktop\git\language_games\steamapps_test"
+#folder_path = "C:\\Program Files (x86)\\Steam\\steamapps"
 config_file_path = "user_language.cfg"
-log_file_path = "start_log.txt"
+blacklist_file_path = "blacklist_appmanifest.cfg"
 
 if not os.path.isfile(config_file_path):
-    with open(config_file_path, "w") as config_file:
+    with open(config_file_path, "w", encoding="utf-8") as config_file:
         config_file.write("english")
 
-log_file = open(log_file_path, "a")
-
-original_stdout = sys.stdout
-sys.stdout = log_file
-
-def log_with_timestamp(message):
-    timestamp = datetime.datetime.now().strftime("%Y-%m-%d %H:%M:%S")
-    log_message = f"[{timestamp}] {message}"
-    print(log_message)
-
-with open(config_file_path, "r") as config_file:
+with open(config_file_path, "r", encoding="utf-8") as config_file:
     config_text = config_file.read().strip()
 
+blacklisted_filenames = set()
+if os.path.isfile(blacklist_file_path):
+    with open(blacklist_file_path, "r", encoding="utf-8") as blacklist_file:
+        for line in blacklist_file:
+            filename = line.strip()
+            if filename:
+                blacklisted_filenames.add(filename)
+
 for filename in os.listdir(folder_path):
-    if filename.startswith("appmanifest_") and filename.endswith(".acf"):
+    if (
+        filename.startswith("appmanifest_")
+        and filename.endswith(".acf")
+        and filename not in blacklisted_filenames
+    ):
         file_path = os.path.join(folder_path, filename)
 
         try:
-            with open(file_path, "r+") as manifest_file:
+            with open(file_path, "r+", encoding="utf-8") as manifest_file:
                 manifest_text = manifest_file.read()
 
-                manifest_text = re.sub(r'("UserConfig"\s*{[\s\S]*?"language"\s*")(.+?)("\s*[\s\S]*?})', fr'\1{config_text}\3', manifest_text, flags=re.DOTALL)
+                manifest_text = re.sub(
+                    r'("UserConfig"\s*{[\s\S]*?"language"\s*")(.+?)("\s*[\s\S]*?})',
+                    fr'\1{config_text}\3',
+                    manifest_text,
+                    flags=re.DOTALL,
+                )
 
                 manifest_file.seek(0)
                 manifest_file.write(manifest_text)
                 manifest_file.truncate()
 
-            log_with_timestamp(f"File {filename} updated.")
+            print(f"File {filename} updated.")
         except FileNotFoundError:
-            log_with_timestamp(f"File {filename} not found. Skipping.")
-
-sys.stdout = original_stdout
-log_file.close()
+            print(f"File {filename} not found. Skipping.")
